Guard against invalid file paths in "Add to Queue" operation
 - Fixes #104345210

+++ Library/LibraryRowItemPart.cs
@@ -137,19 +137,44 @@ namespace MatterHackers.MatterControl.PrintLibrary
 
 		public async override void AddToQueue()
 		{
+			var printItemWrapper = await MakeCopyForQueue();
+
+			// Early exit if MakeCopy failed
+			if(printItemWrapper == null)
+			{
+				return;
+			}
+
 			// create a new item that will be only in the queue
-			QueueData.Instance.AddItem(await MakeCopyForQueue());
+			QueueData.Instance.AddItem(printItemWrapper);
 		}
 
 		private async Task<PrintItemWrapper> MakeCopyForQueue()
 		{
 			var printItemWrapper = await this.GetPrintItemWrapperAsync();
 
+			// Handle non-existing files
+			if(!File.Exists(printItemWrapper.FileLocation))
+			{
+				return null;
+			}
+
 			PrintItem printItemToCopy =  printItemWrapper.PrintItem;
 			string fileName = Path.ChangeExtension(Path.GetRandomFileName(), Path.GetExtension(printItemToCopy.FileLocation));
 			string newFileLocation = Path.Combine(ApplicationDataStorage.Instance.ApplicationLibraryDataPath, fileName);
 
-			File.Copy(printItemToCopy.FileLocation, newFileLocation);
+			// Handle file read/write errors
+			try
+			{
+				File.Copy(printItemToCopy.FileLocation, newFileLocation);
+			}
+			catch(Exception ex)
+			{
+				string errorMessage = string.Format("Unable to duplicate file for queue: {0}\r\n{1}", printItemToCopy.FileLocation, ex.Message);
+				Trace.WriteLine(errorMessage);
+
+				return null;
+			}
 
 			return new PrintItemWrapper(new PrintItem(printItemToCopy.Name, newFileLocation)
 			{
